fix: remainingPoint 자료형 변경
Long -> int

+++ server/Strange505/src/main/java/com/example/Strange505/pointHistory/dto/PointHistoryResponseDto.java
@@ -16,6 +16,6 @@ public class PointHistoryResponseDto {
     private int diff;
     private String description;
     private Timestamp actionDate;
-    private Long remainingPoints;
+    private int remainingPoints;
 
 }
